Dialog moved to a specific component
